Jasmine 2 issue with final done needing to be called if declared as an argument.
A consequence of this fix is that if there are any async blocks among a group
consisting of a Then and one or more Ands, the Then block must be declared as
an async block with a done callback argument. This shouldn't be a problem since
the purpose of the Ands is to deal with idempotent blocks where order ought not
to matter. So any of the async blocks could be the Then block, and all the
others will work as intended.